Load strings and plurals lazily, only when needed
Summary:
This is expected to help reduce startup time/ANR for when a string is needed
quickly but we may not have time to load all strings.

Note that there are a lot of possible optimizations to be done here. But they are kept
for future diffs in order to keep this diff reviewable.

Reviewed By: gauravlochan

Differential Revision: D19636102

fbshipit-source-id: 637bd9a