Revert "Use binary search when finding strings in ParsedStringPack"
Summary:
According to the task T68340959, there is some issue with binary search of string pack ids and we need to bring WhatsApp Android to point to the latest StringPacks open source repo which does not have this issue.
Hence, reverting this Diff, for now, so that other work, like Mmapping can be continued on the repo.

Differential Revision: D22024976

fbshipit-source-id: 9776c0e54a1f5fb07e59da2288b979e097197c92