Replacing SparseArray (not thread-safe) with ConcurrentHashMap (thread-safe)
Summary:
There are a lots of crashes we are seeing in StringPacks which we suspect are because of the racing condition between threads when fetching strings.
Since, we are using `SparseArray` in `ParsedStringPack.java`, which is not thread-safe, we are trying to replace it with `ConcurrentHashMap` which is an alternative to SparseArray with thread safety already considered.

Differential Revision: D22168866

fbshipit-source-id: c5c0005