Fixing the issue with WhatsApp main app not building because it could not find variables defined in gradle.properties
Summary:
D21937920 introduced an issue where the variables defined in `sample/gradle.properties` are not available to `library` when used in main whatsapp app.
Since, whatsapp's app points to `library` folder and it doesn't know anything about the gradle.properties present in `sample/` folder, the main whatsapp app failed the build when I tried to point the main whatsapp app to the tip of master branch of stringpacks submodule.

Differential Revision: D22286601

fbshipit-source-id: 610cbc2