Updated the dependency in sample app to point to library project so that it can work with latest library code
Summary:
I saw that when sample app was depending on maven dependency of the stringpacks library, there was an issue that sample app was not working fine. It was not working fine when sample app was launched and we tried to change the language inside the app.
The reason is that v0.1.0 of the stringpacks library doesn't work with changing the language in the app. If I change the language of the device and relaunch the sample app, it was working fine.
Hence, reverting the sample app to use project library so that we can test it against latest changes in the library.

Differential Revision: D22287767

fbshipit-source-id: d34f025