Fixing the NullPointerException when fetching ClrdResources for a locale that is not supported
Summary:
Recently we pointed whatsapp main app to the tip of master branch of StringPacks. StringPacks' master branch has had lots of improvements over the period of time which were not tested as we were originally pointing to `scratch-whatsapp` branch.
This Diff fixes a crash which happens when the app is trying to get a ClrdResource when a locale is not supported.

If a locale is not supported, it will fallback to US English.

In the next Diff D22756973, I am also adding a unit test in `WhatsAppLocaleTest.java` that will check whether the application works fine for unsupported locales.

Differential Revision: D22749985

fbshipit-source-id: 1851faf