Extract out legacy logic from parsed string pack
Summary:
This extracts out legacy logic for ParsedStringPack so that we can add MMapping in parallel to this.
New class is LoadedStringPack
I also add a boolean to top level getString and getQuantityString methods that would be used with ab props later.

This is coming from D22420009, I seemed to have messed that one up while publishing to remote test branch.

Reviewed By: Ider

Differential Revision: D22490943

fbshipit-source-id: c22afd1