Updating version to 0.2.0
Summary:
We are planning to release new version of StringPacks to open source that will have some bug fixes (but not mmap)
Updating version of the library to 0.2.0
This Diff is on top of `rWASP6b09ba9823f7` commit in StringPacks repo.
I created a local branch off of this commit and created this Diff on top of it.

Reviewed By: Ider

Differential Revision: D25485788

fbshipit-source-id: 481f2a6