Moving the linter engine in to whatsapp/
Summary:
Refactoring `tools/` folder inside `whatsapp/tools/` so that WhatsApp specific files, like lint engine files are not available on GitHub

Also, posted this question on [Open Source - Support](https://fb.workplace.com/groups/osssupport/permalink/4994638037251438/) group about the best way to approach this problem. The basic gist is `Assume damage is already done`. If the files have to be removed (if they are leaking something or for legal reasons): -
- Move the files into one of the filtered folders, and/or
- Re-write the github history as mentioned [here](https://www.internalfb.com/intern/wiki/Open_Source/Internal_Tools/ShipIt/Deleting_history_from_a_ShipIt-synced_repository/)

Disadvantage of rewriting the history (to remove these files from history) would be that: -
- It might raise more eye-brows and bring more attention
- Mess up the setup of anybody who has cloned the repo.

We should rewrite history, only if we HAVE TO (say, for legal reasons). Since, these files are not leaking any confidential information and don't appear to be related to any license violations, we don't have to rewrite the history.

Since, these files are useless to the Open Source project, it is ok to remove them, which this Diff is accomplishing.

Differential Revision: D25934310

fbshipit-source-id: e4141c6