Change StringPacks code generation to support > 4k strings
Summary:
A java method/field has a bytecode size limit of 64K bytes, and this limit is exceeded when we generate the int array for over 4000 strings, causing a compilation error.
Changed to instead generate methods which each have no more than 4000 strings.
Updated templates to reflect the new usage of the generated code.

Reviewed By: ashish29dec

Differential Revision: D25810690

fbshipit-source-id: 6bc3ba5