Removing packIds from StringPackIds.java
Summary:
## Context
Recently it was brought to my attention that developers are facing merge conflicts in StringPackIds.java when they rebase from master.
The reason for this is a combination of a couple of things
- `find` command was recently added to `Makefile` and will run on a daily basis causing the entire `StringPackIds.java` to generate every day. (D26769441), and
- If a developer wants to delete a string, they have to delete the string from `res/values/strings.xml` and also delete its id from `StringPackIds.java`

Hence, if a developer who has a local commit deleting a string, they will run into a merge conflict whenever they rebase with master (if `StringPackIds.java` has changed since last rebase)

## Solution
I created a [quip document](https://fb.quip.com/ph1YA8ma3pfz) discussing the potential solutions

From this document, I am implementing the solution where we generate `StringPackIds.java` in such a way that it doesn't result into conflicts when it is modified.
The main reason why it was resulting into conflicts was because every day `StringPackIds.java` was generated fresh and some strings may have been added or deleted, changing the `packIds` of the `string ids`.
Hence, if a developer has a local copy where they delete a id and master branch's copy has also modified that line (changed the packId), it was resulting into a conflict.
Hence, once we get rid of packIds, we should be fine: -
- Getting rid of `packIds` will enforce the strings mentioned in `StringPackIds.java` to be in sequential order (starting from 0, 1, 2, so on)
- Since `.pack` file is generated as part of gradle task (in WhatsApp) we will always pack the correct `packId` for a string
- In Open Source, whenever someone pulls this commit and wants to move more strings to `StringPackIds.java`, the new script will automatically create new `StringPackIds.java`, move new strings apprpriately and pack the strings with correct packIds.
- The new scripts will not affect Open Source unless someone explicitly runs the `assemble_string_packs.py` command, which will perform all 3 sub-commands correctly (`find`, `move`, `pack`)

Reviewed By: Ider

Differential Revision: D27107492

fbshipit-source-id: 7edb199