Fix app crash when adding a WebView with application Context
Summary:
Fixing the crash reported by Microsoft's Android team when adding a `WebView` in an `Activity` with Application Context

When a `WebView` in added to an `Activity` with `getApplicationContext()` it internally calls `getBaseContext()` which was throwing a crash.
The reason `UnsupportedException` was explicitly being thrown was to discourage people from fetching translations by calling `getBaseContext().getString(int resId)` because it won't fetch the translations from StringPacks

For now, I am removing throwing an Exception. Later I will add a log statement warning the users instead of throwing `UnsupportedException` and then continue to returning `super.getBaseContext()`

Reviewed By: Ider

Differential Revision: D27681544

fbshipit-source-id: 37c135f