Allow for string overrides
Summary:
Previously, string overrides were not supported; if a string's value was overridden in another module or flavor, for example, an exception would be thrown.
Instead, string overrides will now log a warning when encountered. When a given string id is present in more than one entry in original_resources_directories, the last entry's value will be used in the resulting .pack files.

For example, if you have:
original_resources_directories = {
    app/src/main,
    app/src/dev
}

and they both contain a value for R.string.foo, the value for in app/src/dev will be used.

A warning will be printed when this occurs.

Reviewed By: ashish29dec

Differential Revision: D28097993

fbshipit-source-id: 138b90a