Migrating to using maven-publish plugin
Summary:
## Context
Upgrading to using `maven-publish` plugin because WhatsApp Android wants to upgrade AGP to 7.0 to fix lint issues. AGP 7.0 depends on Gradle v7.0 and `maven` plugin has been removed in this version.
Hence, for WhatsApp Android to be able to upgrade to latest Gradle this subrepo needs to be updated

## In this Diff
Upgrading the AGP of StringPacks library to `3.6.0` because starting this version `maven-publish` plugin was supported for Android artifacts (AAR, APK, AAB).
Since AGP v3.6.0 depends on Gradle `v5.6.4` upgrading Gradle as well

Differential Revision: D28550271

fbshipit-source-id: 0e20bc6