Fixing language change from non English language to en-US
Summary:
Microsoft team brought up an issue where changing language from `en-US` to `fr-FR` worked fine for them but changing it back from `fr-FR` to `en-US` didn't and continued to show French translations.
The reason was that we were not setting `parsedStringPack` to null when switching languages.

I also added the functionality to switch between `en-US` and other languages in the sample app.

Reviewed By: Ider

Differential Revision: D29115330

fbshipit-source-id: 7599c99