Ensuring that pack files with same content and in same order are created if no changes have been made
Summary:
When creating incremental builds of the app, i.e. running `./gradlew assembleDebug`, without making any changes to the codebase we see that the order of the data placed in a `.pack` file is not the same.
Ideally, the contents and order in which the contents appear in a `.pack` file should not change if we are not making any changes.

The areas that were not consistent between the runs were the order in which we were storing `pack_ids` and the information about the string it represents and also the strings.

To ensure a consistent order of `pack_ids` and the string's metadata (string's start location and length) and the location of actual string, we are now adding strings to the list in sorted order of pack_ids.
This ensures that a string with `pack_id` 0 will always get processed first and its corresponding string will always get added to the list first.

Differential Revision: D33293913

fbshipit-source-id: 8a57755