Adding error handling in find commands
Summary:
Added error handling in the `find` command in cases where an xml file is not parsable by `ElementTree` API.
We started seeing this crashing the `find` command by some xml files that were ir-relevant for stringpacks. We do not expect this error to be thrown in any of the xml files that we are interested in.

Differential Revision: D33348856

fbshipit-source-id: 0cee4a2