Modernize StringPack with Java 8 Language Feature
Summary:
With all the advancement in Android tooling, desugaring (using d8) can already handle the case more elegantly using [Java 8 feature](https://developer.android.com/studio/write/java8-support-table) of `IntSupplier`.

Hence removing the need of passing in `Object` that need to be type cast everywhere. It is bad for static analysis and confusing enough.

Why we need `IntSupplier` or lazy evaluation? Because evaluating a number (especially in string format) is super costly. If we don't have plural rule at all, then it's better to not do it.

Reviewed By: mdzyuba

Differential Revision: D42795718

fbshipit-source-id: 1a6aacd827dfb1ff515d3f5b46639ea8d5b796b1