Make StringPack use stable order for string.xml files.
Summary:
As
# Background

In Android there is a notion of resource overriding. If two different modules provide the same resource (f.e. string) with different values, only one will win. With Buck we use a topological sorting, which means that if target X depends on target Y, then X will override any values from Y.

# Problem

This resource overriding setup works fine for string resources that are included directly into APK / resources.arsc file. But it behaves randomly for stringpack files. For a few string resources that are (accidentally) overridden in the WhatsApp codebase the value that gets for a particular locale is mostly random - it changes from build to build.

See P1746491362 for a detailed analysis.

# Root cause

Ultimately StringPack does not behave the same as Buck (in current implementation). There are two problems:

1. The ordering of string files is randomized because of using set, which loses order (fixed in this diff)
2. The order of string files is not the same as Buck. So the actual string value in StringPack will be different compared to what we'd include in default US locale (not addressed in this diff)

# Fix

As part of investigation a specific location in the StringPack packing script was identified as the one leading to randomized order of files. Instead of using `set()` this diff uses `list()` while still preserving that only unique file will be added (in practice for WhatsApp Android this shouldn't happen anyway).

Differential Revision: D70600309

fbshipit-source-id: ad5930cd628ed37a07a72aa4ccc5daec13d477f8