Create remap command in WA string pack scripts
Summary:
Creates a new command in `wa_string_pack_command.py` called "remap", which takes in a remap file like
```
10 12
14 13
16 14
```
which tells the script to take an existing string pack and remap all the string IDs inside (10 -> 12, 14 -> 13, 16 -> 14) and write out a new string pack with that remapping. If a string in the old pack does not exist in the mapping it is deleted.

Differential Revision: D86634166

fbshipit-source-id: 077b9ea17914745bab86b7e16f3d0ebdcda3c190